Use CRLF line endings in GET request
The HTTP::Request->as_string() method uses \n as an end-of-line token
default.  Apache HTTPD and Lua's http_server don't tolerate these as
line endings, and throw 400 errors promptly.  Providing CRLF explicitly
as the end-of-line character makes the request palatable to these
servers, and probably others with the same strictness.